Things & stuff
Added dotProperty.js -- A simple "module" that provides support to get/set properties based on their path, and can handle properties that include dots in them. +some tests for it.

Also moved simple-vdf2.js into the repo for now as I have made some bugfix changes to it that need to be kept & updated index.js to point to it rather than using the npm(yarn)-installed verison.